style(assessmentform): abstract yup validation out
